blood-cell-detect/requirements.txt
flask
pyyaml
ultralytics
torch
torchvision
opencv-python
numpy
scikit-learn
kagglehub
matplotlib